+++ seacatauth/session/service.py
@@ -424,7 +424,7 @@ class SessionService(asab.Service):
 		try:
 			await upsertor.execute(event_type=EventTypes.SESSION_EXTENDED)
 		except KeyError:
-			L.warning("Conflict: Session already touched.", struct_data={"sid": session.Session.Id, "v": version})
+			L.error("Conflict: Session already touched.", struct_data={"sid": session.Session.Id, "v": version})
 
 		return await self.get(session.SessionId)