+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 
 ### Fix
 - Add `id_token_signing_alg_values_supported` in OIDC discovery (#260)
+- Add support for `nonce` authorize parameter (#263, PLUM Sprint 230908)
+- Handle special characters in login ident (#264, PLUM Sprint 230908)
 
 ### Features
 - Unified SMTP config (#257, PLUM Sprint 230825)
 - Endpoint for pruning old audit entries (#259, PLUM Sprint 230825)
 - Algorithmic (stateless) anonymous sessions (#232, PLUM Sprint 230825)
+- Initialize Zookeeper service (#265, PLUM Sprint 230908)
 
 ### Refactoring
 - Last login info moved to a dedicated collection (#259, PLUM Sprint 230825)